CMTS parsing and RIP (#407)
* add cisco cable unit test

* parsing and post-processing of cisco cmts cable stanzas

* update unit test bash function

* initialize maps

* cleaner handling of 'rule'

* fix newlines

* cable additions and fixes

* ignore ip spd

* parsing and definition/references for:

- controller
- depi-class
- depi-tunnel
- l2tp-class

* cable/interface parsing

* rip full pipeline implementation, ospf cleanup

* add missing break

* ignore cisco logging cmts

* ignore cisco access-list dynamic-extended

* ignore cisco snmp-server system-shutdown

* ignore cisco depi, packetcable

* ignore cisco linecard-group

* allow 'Modular-Cable' cisco interface name

* allow cisco interface name 'Wideband-Cable'

* ignore cisco isis area-password, update isis tests

* change metric to long to accomodate unsigned int max

* mark acls used in cisco cops listener access-list

* format

* cisco aaa block forms

* ignore cisco tftp-server

* ignore more of cisco monitor

* ignore cisco pseudowire-class control-word

* allow 'global' in cisco ip route

* update refs

+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/AbstractRoute.java
@@ -66,8 +66,8 @@ public abstract class AbstractRoute implements Serializable, Comparable<Abstract
     if (ret != 0) {
       return ret;
     }
-    Integer lhsMetric = getMetric();
-    Integer rhsMetric = rhs.getMetric();
+    Long lhsMetric = getMetric();
+    Long rhsMetric = rhs.getMetric();
     if (lhsMetric == null) {
       if (rhsMetric != null) {
         ret = -1;
@@ -77,7 +77,7 @@ public abstract class AbstractRoute implements Serializable, Comparable<Abstract
     } else if (rhsMetric == null) {
       ret = 1;
     } else {
-      ret = Integer.compare(lhsMetric, rhsMetric);
+      ret = Long.compare(lhsMetric, rhsMetric);
     }
     if (ret != 0) {
       return ret;
@@ -144,7 +144,7 @@ public abstract class AbstractRoute implements Serializable, Comparable<Abstract
     nhip = !Route.UNSET_ROUTE_NEXT_HOP_IP.equals(nextHopIp) ? nextHopIp.toString() : "N/A";
     String net = getNetwork().toString();
     String admin = Integer.toString(getAdministrativeCost());
-    String cost = Integer.toString(getMetric());
+    String cost = Long.toString(getMetric());
     String prot = getProtocol().protocolName();
     String routeStr =
         String.format(
@@ -157,7 +157,7 @@ public abstract class AbstractRoute implements Serializable, Comparable<Abstract
   public abstract int getAdministrativeCost();
 
   @JsonIgnore
-  public abstract Integer getMetric();
+  public abstract Long getMetric();
 
   @JsonProperty(PROP_NETWORK)
   @JsonPropertyDescription("IPV4 network of this route")
